requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
opencv-python==4.10.0.84
numpy>=1.21.0,<2.0.0
scikit-image==0.25.0
pymongo==4.10.1
gunicorn==21.2.0
Pillow>=10.1
scikit-learn==1.2.2
faiss-cpu
